fixed some namings from js to jsx, fixed the search functionality
