+++ src/components/Pagination/Pagination.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import Page from './Page';
+import Page from './Page.jsx';
 
-const Pagination = ({ currentPageNumber, totalNumberOfPages, onChange }) => {
+const pagination = ({ currentPageNumber, totalNumberOfPages, onChange }) => {
   const pages = Array.from(
     Array(totalNumberOfPages).keys()
   ).map((pageNumber) => (
@@ -21,4 +21,4 @@ const Pagination = ({ currentPageNumber, totalNumberOfPages, onChange }) => {
   return <ul className='pagination'>{pages}</ul>;
 };
 
-export default Pagination;
+export default pagination;
