+++ src/data/mockData.json
@@ -65,7 +65,67 @@
       "email": "BertramLBrandt@teleworm.us",
       "edit_path": "http://google.com",
       "per_id": 11
+    },
+    {
+      "name1": "Brad L. gibson",
+      "email": "brad.gibson@example.com",
+      "edit_path": "http://google.com",
+      "per_id": 12
+    },
+    {
+      "name1": "Simon I. Pegg",
+      "email": "simonpegg@you.us",
+      "edit_path": "http://google.com",
+      "per_id": 13
+    },
+    {
+      "name1": "Michale L. Scott",
+      "email": "michalescott@something.com",
+      "edit_path": "http://google.com",
+      "per_id": 14
+    },
+    {
+      "name1": "John L. Snow",
+      "email": "johnSnow@example.it",
+      "edit_path": "http://google.com",
+      "per_id": 15
+    },
+    {
+      "name1": "Jeniffer L. Love",
+      "email": "jenlove@nothing.us",
+      "edit_path": "http://google.com",
+      "per_id": 16
+    },
+    {
+      "name1": "Ingwer L. Brand",
+      "email": "ingwerbrand@blabla.us",
+      "edit_path": "http://google.com",
+      "per_id": 17
+    },
+    {
+      "name1": "Someone L. New",
+      "email": "someonenew@teleworm.us",
+      "edit_path": "http://google.com",
+      "per_id": 18
+    },
+    {
+      "name1": "Bertram L. Feinman",
+      "email": "feinman@physics.us",
+      "edit_path": "http://google.com",
+      "per_id": 19
+    },
+    {
+      "name1": "Samuel L. Jackson",
+      "email": "BertramLBrandt@teleworm.us",
+      "edit_path": "http://google.com",
+      "per_id": 20
+    },
+    {
+      "name1": "Chris L. Hemmingway",
+      "email": "hemmingway@teleworm.us",
+      "edit_path": "http://google.com",
+      "per_id": 21
     }
   ],
-  "total": 11
+  "total": 21
 }
